requirements.txt
streamlit
yfinance
altair
pandas
google-generativeai
numpy
plotly
python-dotenv